fix(homelab): stop rejecting a p95 that is too fast to bucket
A drill with a latency objective could not be finished by the single best move
available in it. Enabling the response cache makes checkout return before it
does any of its per-request CPU work, so requests complete in a few tenths of
a millisecond — inside Envoy's smallest histogram bucket (le=0.5ms). The
scrape rounded that to the integer 0, and the goal check read `p95 > 0` as
"did we get a sample", so it treated the best possible result as a missing
one. The objective sat unmet with the cluster serving 999 of 1000 offered
requests a second at zero errors.

Whether a p95 can be trusted is a question about the sample, not about the
number. It now asks whether traffic is actually completing: if requests are
being served, the latency they were served in is real however small, and a
windowed scrape taken mid-rollout still reports no throughput and so still
fails the check it was written for.

p95 keeps a decimal place through the API and the HUD, because "0.4 ms" is a
result and "0 ms" looks like a broken gauge.

+++ apps/api/Runs/LabRunModels.cs
@@ -86,9 +86,11 @@ public sealed record RunTelemetry(
     int PostgresCpuPct,
     int ApiReplicas,
     bool CacheEnabled,
-    // Real request metrics scraped from the run's Envoy gateways.
+    // Real request metrics scraped from the run's Envoy gateways. p95 carries a decimal place
+    // because a cached request is genuinely sub-millisecond and whole milliseconds cannot tell
+    // "0.4ms" from "we did not get a sample".
     int RequestsPerSec,
-    int P95LatencyMs,
+    double P95LatencyMs,
     double ErrorRatePct);
 
 // One pod of a cluster component, with its measured resource usage.
